requirements.txt
mysql-connector-python
pandas
reportlab
python-docx
qrcode
matplotlib
beautifulsoup4
pyttsx3